initial tuple support
implemented as anonymous structs.
the syntax is currently (x, y) however that causes huge complexity
using a hackish lookahead to skip over expressions without parsing them
and rewinding the parser (in is_tuple...) and will be changed to {x, y} -
which is coincidentally in line with the planned syntax for struct
initializers (and later array initializers).