foundation for per-var-reference tracking
currently refs are saved as a bitflag on the type_id, which is quite
hackish. with the about-to-be-added byref keyword for function parameters,
it's better to track per-variable whether something's a reference,
and also what kind - stack or heap. stack structures passed to a function
with byref parameter will automatically be passed as a reference.