use handle_assignment also from led()
this automatically gives us all the goodies from the central
handler for statement expressions (e.g. in if or while conditionals).
the only difference is that in statement expressions, compound assign
literals are not accepted (yet?).
there's currently also no code to deal with function call result
assignments, unlike in statement().