implement default initializers and nil, constructors on arrays of struct
previously, a variable declaration like `var x:int` on its own wasnt
allowed; it always had to have an explicit value assigned.
this is now allowed, with an explicit 0 assigned for it.
float types get 0.0, strings "", and structs etc are zero-initialized.

we now also allow `new` combined with an array type, e.g. var x:=new int[3].
in order to allow both constructor and non-constructor use for the array
use, var p:= new Point[2] does not call the constructor, but initializes
all fields to 0. var q := new Point[2]() otoh runs the (default) constructor
for each element. this of course only makes sense if the arguments (if any)
of the constructor are identical for each item.

nil is introduced for variables with by-ref context where we dont want
to allocate the storage just yet - the plan is to add a parenthesized
version of new() that can reallocate. the other use case is to check
for whether memory allocation failed.