add a token to each AST node
initially, the plan was to use e.g. for binary and unary ops the
token.type instead of a string like 'bitnot' - the token's string
value used in the constructor, but it's not that straightforward
as we often use a different value than what's actually in the
token. additionally, the token that triggered a specific codepath
is often advanced many times until Node creation time.
however, having the token information in the node is still a boon
because it allows us to raise errors from the interpreter with
somewhat accurate source line/col information.

the interpreter side using the token for the error isnt implemented
yet.