replace StructInit node with CallNode/GenericInitializer...
and add hidden self return to constructors.

the StructInitNode was used to signify

- construct an object instance
- call constructor if it exists, else zero-init the struct instance

this node was always a thorn in my eye since we had to duplicate
basically all of the regular CallNode handling, and it became
especially painful in the C codegen, because the language semantics
dictate that for "syntactic sugar" reasons, constructors are
written as void functions, but actually behave like:

- (storage is allocated before constructor call)
- call constructor with hidden first byref self param pointing to that storage
- return self (again byref) so calls can be chained.

so we'd always encounter this call, in potentially complex expressions,
behaving like it would return a value, but actually didn't.

in order to get the semantics right, we now:
- modify constructors in-flight to add a return type equivalent to self
- turn every "return" into "return self"
- add an explicit "return self" at the end.

this finally gives the "internal" constructor all of the semantic
properties, so it can be used (almost) like any other function call.
almost, because CallNode expects to be used with an already existing
VariableNode of the struct type. That's why i had to patch in a
special __dunno__ variable node that needs to be replaced with
the instance creation at interpreter/codegen side.