fix constructor return semantics to be either byval or byref
the constructor is special in that if it's combined with "new",
it returns a struct ref, else a stack object by value.
at least as far as the semantics of the language are concerned.
the C codegen actually needs to turn both cases into byref
semantics, so we can use safe handle-allocation for all objects,
but we'll implement that in a separate change.