ccodegen: properly dereference variables encountered in initializers
when we encounter an initializer that uses a variable instance, e.g.:

    var p := Point()           # p becomes a handle (ref) in the C codegen
    var t := {p, "coordinate"} # t becomes a tuple of { Point, string }

or

    t = {p, "coordinate"}

p will be of handle type because we turn all struct instances into
reference types for smooth memory management in the C codegen; however
tuples are always value types, so if we encounter a struct instance
variable in the initializer, we need to dereference it.

in order to support the full range of nested expressions we could
encounter in an initializer, this requires recursive tracking of
what item in a compound type we currently act upon, which this commit
implements.

fortunately, in the parser, the initializer gets assigned the type of
the LHS, which lets us omit special case handling in var_decl and binop
with = operator to "set" what the recursive parsing of the RHS should
act upon.

an alternative to implementing this initializer context stack would be
an AST walker pass that either completely decomposes initializers into
separate field assignment statements (but this would be very verbose
on the emitted C code and could considerably bloat both the code and
the binary produced), or sets a flag upon encountering VariableNode
instances of struct types in an initializer that this item needs to be
dereferenced.