ccodegen: initial stab at implementing generics
this works mostly, but generics are not that useful without
dynamic array resizing, which we don't handle yet.
additionally, the point at which we emit the concrete struct
definitions might not be correct under all circumstances, but
works for the existing testcases.

in the worst case, we might have to do a full dependency graph
for all types, structs, functions, etc to emit everything in
the right order.