ccodegen: change injected handle ref_kind to stack, support byval array params
we always use handles for arrays and structs, which need to be of a
byref type, however this produces a dilemma when those values are
actually needed in a by-value context.
since REF_KIND_STACK was unused, and _GENERIC used for explicit byref
function args and return values, we now assign the stack ref_kind so
we can "dereference" handles with injected ref_kind that are semantically
by-value.

using this, we fix byval array parameters by renaming the parameter,
setting its ref_kind to byref, and inject a new vardecl with the
original parameter name that's by-value and assign the byref param
to it. that allows to have the storage of such arrays on the stack,
which is semantically identical as if it was actually passed by
value as a C function parameter (for arrays, those would need to
be wrapped in a struct because arrays decay to pointers).

the function_context param added to some ast_flattener functions
means that the expr is a function call param or return value,
so it can inject the right REF_KIND_GENERIC if needed, because
REF_KIND_STACK would be wrong for those.