ccodegen: major refactoring: use handles only for byref vars
it turned out that "use handles for structs and arrays, regardless
of whether they're byval or byref" caused way too much complications,
mostly because handles always have reference semantics.
the final nail in the coffin was the problematics of modeling this
in "arrays in structs".

now, we just behave like the language semantics. if it's by-value,
then we have a "struct something", else "handle".
note that we now wrap all non-dynamic array types in a struct so
we can pass them by-value to C functions without having to inject
temporary objects copied from a pointer arg. (in C arrays decay to
pointers).
also a new handle allocator type had to be added for global
variables.
in general, we now always create a helper handle for any by-value
struct/array type, so we already have it in case we need to pass
a handle as a byref function arg.

we currently pass 84.5% of the testsuite, a minor regression vs the
previous implementation which passed 85.5%.
most of the failing testcases use either inheritance or strings,
both of which arent implemented yet.

the code isn't particularly cleaned up yet, but so many changes
accumulated that it's about time for a checkpoint in git.